python310 minimal and always download pure-python wheels (#4986)
* installer.functions: New install_python_wheels() function helper

* wheel.mk: Allow enabling pure-python compiling (default disabled)

* homeassistant: Migrate to using new python wheel install helper

* python310: Use new python wheel installation helper

* homeassistant: Move requirements back from python310

* homeassistant: Use generic wheel installer

* beet: Move requirements back from python310

* beet: Use generic wheel installer

* python310: Remove all non-mandatory wheels (e.g. basic package)

* wheel.mk: noarch package may not have any *.whl files

* installer.functions: Add a virtualenv default function

* installer.functions: Allow dependency mode for pure python wheels

* beets: Adjustments to requirements.txt

* python310: Adjustments to requirements for pillow & beets

* homeassistant: Adjustments to requirement files

* installer.functions: Accept parameters to installing wheels

* sickchill: Migrate to using function based wheel installer

* sickchill: Final adjustments to requirements

* mercurial: Update to using standard wheel functions

* borgbackup: Migrate to using standard wheel install functions

* duplicity: Use standard method to create usr/local/bin symlink

* python310: Simplify requirements to bare minimal

* borgbackup: Re-adjust to new python310 requirements

* duplicity: Adjust to new python310 requirements

* homeassistant: Adjust to new python310 requirements

* sickchill: Adjust to new python310 requirements

* duplicity: Update lxml and enforce pbr versions

* rdiff-backup: Migrate to using standard wheel install method

* python310: Reducing package wheels to bare minimal

* sickchill: Use alternate tmp at installation

* sickchill: Move back compiled wheels from python310

* borgbackup: Update packaging wheel

* beets: Move to arch-specific (i.e. no longer noarch)

* octoprint: Migrate wheels back from python310 and use default inst.

* duplicity: Migrate back wheels from python310

* homeassistant: Move back cross-compiled wheels from python310

* installer.functions: Add method to install all wheels at once

* sickchill: Rename requirements-cross.txt to crossenv

* sabnzbd: Migrate to using standard wheel install function

* flexget: Update to version 3.1.155 and use new installer funct.

* python310: Bump package version

* installer.functions: Remove alt wheel installer function

* flexget: Fix for arm user alt install dir and add missing wheels

Was missing cryptography and cffi in order to install properly
on arm (tested on armv7)

+++ mk/spksrc.wheel-env.mk
@@ -2,6 +2,11 @@
 # Configuration for python wheel build
 #
 
+# Enable pure-python packaging
+ifeq ($(strip $(WHEELS_PURE_PYTHON_PACKAGING_ENABLE)),)
+WHEELS_PURE_PYTHON_PACKAGING_ENABLE = FALSE
+endif
+
 ifeq ($(strip $(WHEELS_DEFAULT)),)
 WHEELS_DEFAULT = requirements.txt
 endif
@@ -73,22 +78,25 @@ install_python_wheel:
 		cd $(WHEELHOUSE) ; \
 		$(MSG) Copying $(WHEELS_DEFAULT) wheelhouse ; \
 		if stat -t requirements*.txt >/dev/null 2>&1; then \
+			cp requirements*.txt $(STAGING_INSTALL_WHEELHOUSE) ; \
 			cat requirements*.txt >> $(STAGING_INSTALL_WHEELHOUSE)/$(WHEELS_DEFAULT) ; \
 			sort -u -o $(STAGING_INSTALL_WHEELHOUSE)/$(WHEELS_DEFAULT) $(STAGING_INSTALL_WHEELHOUSE)/$(WHEELS_DEFAULT) ; \
 		fi ; \
-		if [ "$(EXCLUDE_PURE_PYTHON_WHEELS)" = "yes" ] ; then \
-			echo "Pure python wheels are excluded from the package wheelhouse." ; \
-			for w in *.whl; do \
-				if echo $${w} | grep -viq "-none-any\.whl" ; then \
-					cp -f $$w $(STAGING_INSTALL_WHEELHOUSE)/`echo $$w | cut -d"-" -f -3`-none-any.whl ; \
-				fi ; \
-			done ; \
-		else \
-			for w in *.whl; do \
-				_new_name=$$(echo $$w | sed -E "s/(.*-).*(linux_).*(\.whl)/\1\2$(PYTHON_ARCH)\3/") ; \
-				$(MSG) Copying to wheelhouse: $$_new_name ; \
-				cp -f $$w $(STAGING_INSTALL_WHEELHOUSE)/$$_new_name ; \
-			done ; \
+		if stat -t *.whl >/dev/null 2>&1; then \
+			if [ "$(EXCLUDE_PURE_PYTHON_WHEELS)" = "yes" ] ; then \
+				echo "Pure python wheels are excluded from the package wheelhouse." ; \
+				for w in *.whl; do \
+					if echo $${w} | grep -viq "-none-any\.whl" ; then \
+						cp -f $$w $(STAGING_INSTALL_WHEELHOUSE)/`echo $$w | cut -d"-" -f -3`-none-any.whl ; \
+					fi ; \
+				done ; \
+			else \
+				for w in *.whl; do \
+					_new_name=$$(echo $$w | sed -E "s/(.*-).*(linux_).*(\.whl)/\1\2$(PYTHON_ARCH)\3/") ; \
+					$(MSG) Copying to wheelhouse: $$_new_name ; \
+					cp -f $$w $(STAGING_INSTALL_WHEELHOUSE)/$$_new_name ; \
+				done ; \
+			fi ; \
 		fi ; \
 	fi
